recruitment_analyzer/requirements.txt
Django==5.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
transformers==4.30.2
torch==2.0.1
numpy==1.24.2
pandas==2.0.2
python-dotenv==1.0.0
requests==2.28.2

sympy~=1.13.3